Expense Tile Design Changes
[Changes]

Made the Design more elegant
Created HomePage for expenses
Worked DateTime Utilities
