+++ lib/UserInterface/Theme/AppTheme.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static const Color textColor = Colors.white;
+  static const Color secondaryTextColor = Color.fromARGB(255, 118, 118, 118);
+  static const Color themeColor = Color.fromARGB(255, 41, 45, 50);
+}
