+++ lib/Utils/logger_utils.dart
@@ -1,0 +1,3 @@
+import 'package:logger/logger.dart';
+
+final Logger logger = Logger();
