+++ lib/Utils/sms_utils.dart
@@ -1,3 +1,4 @@
+import 'package:expense_tracker/Utils/logger_utils.dart';
 import 'package:flutter_sms_inbox/flutter_sms_inbox.dart';
 import 'package:permission_handler/permission_handler.dart';
 
@@ -5,7 +6,7 @@ class SMSUtils {
   static Future<List<SmsMessage>> getSMS() async {
     PermissionStatus status = await Permission.sms.request();
     if (status.isDenied) {
-      // TODO: add a logger for denied status.
+      logger.e('sms permission denied.');
       return [];
     }
 
@@ -15,12 +16,12 @@ class SMSUtils {
         SmsQueryKind.inbox,
       ],
     );
-    print('sms inbox messages: ${messages.length}');
+    logger.i('sms inbox messages: ${messages.length}');
     return messages;
   }
 
   static Future<void> saveSMSInJSON() async {
     final messages = await getSMS();
-    print('saving sms inbox messages: ${messages.length}');
+    logger.i('saving sms inbox messages: ${messages.length}');
   }
 }
