+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   flutter_sms_inbox: ^1.0.3
   permission_handler: ^10.4.3
   path_provider: ^2.1.0
+  logger: ^2.0.1
 
 dev_dependencies:
   flutter_test:
